+++ src/unpoly/classes/compiler_pass.js
@@ -3,22 +3,24 @@ const e = up.element
 
 up.CompilerPass = class CompilerPass {
 
-  constructor(root, compilers, options = {}) {
+  constructor(root, compilers, { skip, layer, data, dataMap } = {}) {
     this.root = root
     this.compilers = compilers
 
     // Exclude all elements that are descendants of the subtrees we want to keep.
     // The exclusion process is very expensive (in one case compiling 100 slements
-    // took 1.5s because of this). That's why we only do it if (1) options.skipSubtrees
+    // took 1.5s because of this). That's why we only do it if (1) options.skip
     // was given and (2) there is an [up-keep] element in root.
-    this.skipSubtrees = options.skip
-    if (!this.skipSubtrees.length || !this.root.querySelector('[up-keep]')) {
-      this.skipSubtrees = undefined
+    if (skip?.length && this.root.querySelector('[up-keep]')) {
+      this.skip = skip
     }
 
     // (1) If a caller has already looked up the layer we don't want to look it up again.
     // (2) Ddefault to the current layer in case the user manually compiles a detached element.
-    this.layer = options.layer || up.layer.get(this.root) || up.layer.current
+    this.layer = layer || up.layer.get(this.root) || up.layer.current
+
+    this.data = data
+    this.dataMap = dataMap
 
     this.errors = []
   }
@@ -28,13 +30,31 @@ up.CompilerPass = class CompilerPass {
     // If we're compiling a fragment in a background layer, we want
     // up.layer.current to resolve to that background layer, not the front layer.
     this.layer.asCurrent(() => {
+
+      this.setCompileData()
+
       for (let compiler of this.compilers) {
         this.runCompiler(compiler)
       }
+
     })
 
     if (this.errors.length) {
       throw new up.CannotCompile('Errors while compiling', { errors: this.errors })
+    }
+  }
+
+  setCompileData() {
+    if (this.data) {
+      this.root.upCompileData = this.data
+    }
+
+    if (this.dataMap) {
+      for (let selector in this.dataMap) {
+        for (let match of this.select(selector)) {
+          match.upCompileData = this.dataMap[selector]
+        }
+      }
     }
   }
 
@@ -114,7 +134,7 @@ up.CompilerPass = class CompilerPass {
 
   select(selector) {
     let matches = e.subtree(this.root, u.evalOption(selector))
-    if (this.skipSubtrees) {
+    if (this.skip) {
       matches = u.reject(matches, (match) => this.isInSkippedSubtree(match))
     }
     return matches
@@ -122,7 +142,7 @@ up.CompilerPass = class CompilerPass {
 
   isInSkippedSubtree(element) {
     let parent
-    if (u.contains(this.skipSubtrees, element)) {
+    if (u.contains(this.skip, element)) {
       return true
     } else if ((parent = element.parentElement)) {
       return this.isInSkippedSubtree(parent)